backend/requirements-test.txt
certifi==2025.11.12
charset-normalizer==3.4.4
coverage==7.13.0
djangorestframework==3.16.1
execnet==2.1.2
factory_boy==3.3.3
Faker==39.0.0
freezegun==1.5.5
idna==3.11
iniconfig==2.3.0
model-bakery==1.15.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
pytest-mock==3.15.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
responses==0.25.8
six==1.17.0
testcontainers==4.13.3
tzdata==2025.3
urllib3==2.6.2
